DOCKER-311 Raise requirements for chown in 90-init-alfresco.sh
Recent experience (USGBE) has pointed out that mounting a large, existing
alf_data with incorrect permissions/ownership resulted in a very
extensive chown. The prefered solution would have been to fix the issue
through changes to the users/groups on the host, or even to the user
set in the image/container.

This change adds an additional check on permissions, and a check on
whether the alf_data folder is too big to be effeciently chowned. The
size check is set to 10000 items in the dir by default, but can be changed through
an environment variable.